+++ src/ts/components/Component.test.ts
