+++ src/ts/components/MapComponent.test.ts
