+++ src/ts/components/MetricComponent.test.ts
