+++ src/ts/libs/Domain.ts
